+++ example/lib/main.dart
@@ -11,9 +11,9 @@ void main() {
     new ndef.UriRecord.fromUriString("https://github.com")
   ];
 
-  // decode full ndef message (concatenation of records)
-  var decodedurlRecords = ndef
-      .decodeRawNdefMessage(ndef.ByteUtils.hexStringToBytes(encodedUrlRecord));
+  /// decode full ndef message (concatenation of records)
+  /// note that we have implemented extension methods on [Uint8List], [String], [int] and [BigInt]
+  var decodedurlRecords = ndef.decodeRawNdefMessage(encodedUrlRecord.toBytes());
 
   assert(urlRecords.length == decodedurlRecords.length);
 
@@ -28,7 +28,7 @@ void main() {
   // modify the record by data-binding
   var origPayload = urlRecords[0].payload;
   print('===================');
-  print('original payload: ' + ndef.ByteUtils.bytesToHexString(origPayload));
+  print('original payload: ' + origPayload.toHexString());
   print('original uri: ' + urlRecords[0].uri.toString());
 
   // change uri
@@ -36,22 +36,19 @@ void main() {
   urlRecords[0].uriData =
       'github.com/nfcim/flutter_nfc_kit'; // thats also our awesome library, check it out!
   print('payload after change uriData: ' +
-      ndef.ByteUtils.bytesToHexString(
-          urlRecords[0].payload)); // encoded when invoking
+      urlRecords[0].payload.toHexString()); // encoded when invoking
   print('uri after change uriData: ' + urlRecords[0].uri.toString());
 
   // change it back (by using payload)
   print('===================');
   urlRecords[0].payload = origPayload; // decoded when invoking
-  print('payload after changed back: ' +
-      ndef.ByteUtils.bytesToHexString(urlRecords[0].payload));
+  print('payload after changed back: ' + urlRecords[0].payload.toHexString());
   print('uri after changed back: ' + urlRecords[0].uri.toString());
 
   // encoded into message again (also canonicalize MB & MF fields)
   var encodedAgain = ndef.encodeNdefMessage(urlRecords);
-  assert(ndef.ByteUtils.bytesToHexString(encodedAgain) == encodedUrlRecord);
-  print('encoded single record: ' +
-      ndef.ByteUtils.bytesToHexString(urlRecords[0].encode()));
+  assert(encodedAgain.toHexString() == encodedUrlRecord);
+  print('encoded single record: ' + urlRecords[0].encode().toHexString());
 
   // also you can decode by providing id, type and payload separately (normally from phone API)
   print('===================');