+++ test/ndef_test.dart
@@ -9,8 +9,7 @@ import 'package:ndef/utilities.dart';
 
 void testParse(List<String> hexStrings, List<List<NDEFRecord>> messages) {
   for (int i = 0; i < hexStrings.length; i++) {
-    var decoded =
-        decodeRawNdefMessage(ByteUtils.hexStringToBytes(hexStrings[i]));
+    var decoded = decodeRawNdefMessage(hexStrings[i].toBytes());
     assert(decoded.length == messages[i].length);
     for (int j = 0; j < decoded.length; j++) {
       assert(decoded[j].isEqual(messages[i][j]));
@@ -20,10 +19,7 @@ void testParse(List<String> hexStrings, List<List<NDEFRecord>> messages) {
 
 void testGenerate(List<String> hexStrings, List<List<NDEFRecord>> messages) {
   for (int i = 0; i < hexStrings.length; i++) {
-    print('!!!!!!');
-    var genString = ByteUtils.bytesToHexString(encodeNdefMessage(messages[i]));
-    print(genString);
-    assert(genString == hexStrings[i]);
+    assert(encodeNdefMessage(messages[i]).toHexString() == hexStrings[i]);
   }
 }
 
@@ -51,12 +47,15 @@ void main() {
   test('ndef message with text type', () {
     List<String> hexStrings = [
       "d1010f5402656e48656c6c6f20576f726c6421",
-      //"d101145485656d6f6a69fffe3dd801de3dd802de3ed828dd"  // can only decode emoji correctly, but can not encode
+      "d101145485656d6f6a69fffe3dd801de3dd802de3ed828dd"
     ];
 
     List<List<NDEFRecord>> messages = [
       [new TextRecord(language: 'en', text: 'Hello World!')],
-      //[new TextRecord(language:'emoji',text:'😁😂🤨',encoding:TextEncoding.UTF16)],
+      [
+        new TextRecord(
+            language: 'emoji', text: '😁😂🤨', encoding: TextEncoding.UTF16)
+      ],
     ];
 
     // parse
@@ -249,7 +248,7 @@ void main() {
               EIRType.ManufacturerSpecificData: Uint8List.fromList([97])
             })
       ],
-      
+
     ];
 
     // parse